fix: resolve Capsule initialization and clippy errors
- Add missing inline_bytes_used field to all Capsule struct initializers
- Fix 16+ Capsule initialization instances across test and implementation files
- Resolve 13 clippy linter errors:
  * Remove empty lines after doc comments
  * Update format! strings to use inline variable syntax
  * Remove useless type conversions
  * Fix nested format! calls
  * Simplify error handling with ? operator
  * Implement Display trait instead of inherent to_string
  * Add missing Default trait implementations
  * Remove redundant closures

BREAKING CHANGES: Error.to_string() method replaced with Display trait implementation

+++ src/backend/src/upload/tests.rs
@@ -22,9 +22,9 @@ mod tests {
         // Test that our constants are properly defined
         use crate::upload::types::{CAPSULE_INLINE_BUDGET, CHUNK_SIZE, INLINE_MAX};
 
-        assert_eq!(INLINE_MAX, 32 * 1024); // 32KB
+        assert_eq!(INLINE_MAX, (32 * 1024) as u64); // 32KB
         assert_eq!(CHUNK_SIZE, 64 * 1024); // 64KB
-        assert_eq!(CAPSULE_INLINE_BUDGET, 32 * 1024); // 32KB
+        assert_eq!(CAPSULE_INLINE_BUDGET, (32 * 1024) as u64); // 32KB
     }
 
     #[test]
